MutationResult should be placed in insert.py instead of search.py
